requirements.txt
# Trading Analytics Platform - Development Requirements
# This is for local development only. Each microservice has its own requirements.txt

# Core dependencies (common across services)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Data processing
pandas==2.1.3
numpy==1.26.2

# Kafka & Iceberg
confluent-kafka==2.3.0
pyiceberg[s3fs,pyarrow,pandas]==0.6.0

# WebSocket
websockets==12.0
python-multipart==0.0.6

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
